chore(xlsx): extract chart walls (3D view, floor, side/back walls) into chart/walls
Move parseView3D / parseFloorThickness / parseSideWallThickness /
parseBackWallThickness / buildView3D / buildFloorThickness /
buildSideWallThickness / buildBackWallThickness / clampView3DInt and
the matching resolveView3D / resolveFloorThickness /
resolveSideWallThickness / resolveBackWallThickness helpers from the
three chart-{reader,writer,clone}.ts monoliths into a single per-host
module. Pure relocation — JSDoc preserved verbatim, no behavior change.

Add chart/util.ts with the shared findChild / findDescendant /
childElements / collectTextRuns / elementText / parseBoolAttr /
readBoolAttr / readBoolVal / formulaText / parseNumericChildVal
helpers so future per-host modules can reuse them without duplicating
the XML walk primitives.
