+++ src/xlsx/chart/index.ts
@@ -12,3 +12,5 @@ export type * from "./types";
 export * from "./shape";
 export * from "./text";
 export * from "./layout";
+export * from "./util";
+export * from "./walls";
